backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
pytz==2024.2
Werkzeug==3.0.3
python-socketio==5.11.4
azure-storage-blob==12.23.1
email-validator==2.1.1
python-multipart==0.0.7
httpx==0.27.0
python-jose[cryptography]==3.3.0
requests==2.31.0
